# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação API - Teams",
        "description": "Microservice Teams",
        "version": "1.0.0"
    },
    "paths": {
        "/api/coach": {
            "get": {
                "tags": [
                    "Coaches"
                ],
                "summary": "Get list of coaches",
                "operationId": "getCoaches",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Coach"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Coaches"
                ],
                "summary": "Create a new coach",
                "operationId": "createCoach",
                "requestBody": {
                    "description": "Coach object to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Coach"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Coach created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Coach"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/api/coach/{id}": {
            "get": {
                "tags": [
                    "Coaches"
                ],
                "summary": "Find coach by ID",
                "operationId": "getCoachById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of coach to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Coach"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Coach not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Coaches"
                ],
                "summary": "Update an existing coach",
                "operationId": "updateCoach",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of coach to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Coach object to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Coach"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Coach updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Coach"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Coach not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Coaches"
                ],
                "summary": "Delete a coach",
                "operationId": "deleteCoach",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of coach to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Coach marked as deleted successfully"
                    },
                    "404": {
                        "description": "Coach not found"
                    }
                }
            }
        },
        "/api/team": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Get list of teams",
                "operationId": "getTeams",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Team"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Teams"
                ],
                "summary": "Create a new team",
                "operationId": "createTeam",
                "requestBody": {
                    "description": "Team object to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Team"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Team created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        },
        "/api/team/{id}": {
            "get": {
                "tags": [
                    "Teams"
                ],
                "summary": "Find team by ID",
                "operationId": "getTeamById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of team to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Team not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Teams"
                ],
                "summary": "Update an existing team",
                "operationId": "updateTeam",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of team to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Team object to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Team"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Team updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Team"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Team not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Teams"
                ],
                "summary": "Delete a team",
                "operationId": "deleteTeam",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of team to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Team deleted successfully"
                    },
                    "404": {
                        "description": "Team not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Coach": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "dob": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "Team": {
                "required": [
                    "coach_id",
                    "name_club"
                ],
                "properties": {
                    "coach_id": {
                        "type": "integer"
                    },
                    "popular_name": {
                        "type": "string"
                    },
                    "nickname_club": {
                        "type": "string"
                    },
                    "name_club": {
                        "type": "string"
                    },
                    "acronym_club": {
                        "type": "string"
                    },
                    "shield_club": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}